requirements.txt
langchain
langchain-core
langchain-huggingface

langchain-openai
openai

python-dotenv

typing

pydantic
pydantic[email]

streamlit